+++ src/Building.h
@@ -43,7 +43,6 @@ public:
 
     virtual void   calc_elevation() = 0;
     virtual void   reconstruct() = 0;
-    virtual void   reconstruct_flat_terrain() = 0;
     virtual void   insert_terrain_point(const Point_3& pt) = 0;
 
     double get_elevation();
@@ -52,7 +51,6 @@ public:
     void   clip_bottom(const TerrainPtr& terrain);
     void   refine();
     void   alpha_wrap(double relativeAlpha, double relativeOffset);
-    void   translate_footprint(const double h);
     bool   is_part_of(const Polygon_2& otherPoly) const;
     void   set_reconstruction_rules(const BoundingRegion& reconRegion);
     void   remove_reconstruction_rules();
@@ -62,7 +60,7 @@ public:
     void   mark_as_failed();
     bool   has_failed_to_reconstruct() const;
     bool   has_self_intersections() const;
-    void   set_to_zero_terrain();
+    void   set_to_flat_terrain(const double elevation = 0.);
     double sq_max_dim();
     PointSet3Ptr get_points() const;
     std::string get_lod() const;
@@ -80,6 +78,8 @@ protected:
     bool                 m_hasFailed;
     bool                 m_clipBottom = Config::get().clip;
     std::shared_ptr<const Config::ReconRegion> m_reconSettings;
+
+    void force_building_terrain_intersection(const double h = 10.);
 };
 
 //-- Struct for clipping