+++ src/Map3d.cpp
@@ -82,11 +82,11 @@ void Map3d::reconstruct() {
         this->set_footprint_elevation(m_surfaceLayersPtr);
     }
 
-    //-- Clip building bottoms
-    if (Config::get().clip) this->clip_buildings();
-
     //-- Flatten terrain with flag
     if (Config::get().flatTerrain) this->reconstruct_with_flat_terrain();
+
+    //-- Clip building bottoms
+    if (Config::get().clip) this->clip_buildings();
 
     //-- Constrain features, generate terrain mesh from CDT
     this->reconstruct_terrain();
@@ -175,8 +175,8 @@ void Map3d::set_features() {
 
     //-- Set flat terrain or random thin terrain points
     if (m_pointCloud.get_terrain().empty()) {
-        m_pointCloud.create_flat_terrain(m_allFeaturesPtr);
-        Config::get().flatTerrain = false; // all points are already at 0 elevation
+        m_pointCloud.create_flat_terrain(m_allFeaturesPtr, Config::get().flatTerrainElevation);
+        Config::get().flatTerrain = false; // all points are already at flat terrain elevation
     } else {
         m_pointCloud.random_thin_pts();
     }
@@ -375,10 +375,12 @@ void Map3d::reconstruct_buildings() {
     }
     int count = 0;
     int tenPercent = m_buildingsPtr.size() / 10;
+    if (tenPercent == 0) tenPercent = 1;
     #pragma omp parallel for
     for (int i = 0; i < m_buildingsPtr.size(); ++i) {
     //for (auto& f : m_buildingsPtr) { // MSVC doesn't like range loops with OMP
         auto& b = m_buildingsPtr[i];
+        if (Config::get().flatTerrain) b->set_to_flat_terrain(Config::get().flatTerrainElevation);
         if (b->is_active()) this->reconstruct_one_building(b);
 
         if ((count % tenPercent) == 0)
@@ -466,15 +468,14 @@ void Map3d::reconstruct_boundaries() {
 void Map3d::reconstruct_with_flat_terrain() {
     //-- Account for zero terrain height of surface layers
     for (auto& sl : m_surfaceLayersPtr) {
-        sl->set_zero_borders();
-    }
-    //-- Account for zero terrain height of buildings
-    for (auto& b : m_buildingsPtr) {
-        if (!b->is_active()) continue; // skip failed reconstructions
-        b->set_to_zero_terrain();
+        sl->set_flat_borders(Config::get().flatTerrainElevation);
     }
     //-- Set terrain point cloud to zero height
-    m_pointCloud.set_flat_terrain();
+    m_pointCloud.set_flat_terrain(Config::get().flatTerrainElevation);
+
+    Boundary::set_flat_borders(Config::get().flatTerrainElevation);
+
+    //-- Buildings are already reconstructed with defined elevation
 }
 
 void Map3d::solve_building_conflicts() {
@@ -549,14 +550,10 @@ void Map3d::wrap() {
 }
 
 void Map3d::read_data() {
-    //-- Read point clouds
-    m_pointCloud.read_point_clouds();
-
     //-- Read building polygons
     if (!Config::get().gisdata.empty()) {
-        std::cout << "Reading polygons" << std::endl;
         IO::read_polygons(Config::get().gisdata, m_polygonsBuildings, &Config::get().crsInfo);
-        if (m_polygonsBuildings.empty()) throw std::invalid_argument("Didn't find any building polygons!");
+        if (m_polygonsBuildings.empty()) throw city4cfd_error("Didn't find any building polygons!");
     }
     //-- Read surface layer polygons
     for (auto& topoLayer: Config::get().topoLayers) {
@@ -582,6 +579,9 @@ void Map3d::read_data() {
                                                                   ".ply. .off, or .json (CityJSON)"));
         }
     }
+
+    //-- Read point clouds
+    m_pointCloud.read_point_clouds();
 }
 
 void Map3d::output() {