+++ src/PolyFeature.h
@@ -53,7 +53,7 @@ public:
                                         std::vector<EPECK::Segment_3>& constrainedEdges,
                                         std::vector<std::pair<Polygon_with_holes_2, int>>& newPolys,
                                         bool& isNextToBuilding);
-    void  set_zero_borders();
+    void  set_flat_borders(const double elevation = 0.);
     void  calc_min_bbox();
     void  clear_feature();